Add Checks For Login/Valid User
Added checks to determine if the login info was valid and if the
user in question is private even after proper login.

Removed threading for writing as it wasn't really doing much nor
properly implemented. Most of the duration of the program is actually
due to selenium and attempting to navigate the page fully.